demo/requirements.txt
Django~=4.2.0
daphne~=4.0.0
psycopg2~=2.9.6
python-decouple~=3.8.0
-e git+ssh://git@github.com/bossan/openpki#egg=open-pki
